+++ sample-bikeshop/src/main/java/com/example/owensbikes/DialogComposer.java
@@ -13,12 +13,6 @@ import com.squareup.sdk.register.RegisterClient;
 public class DialogComposer {
 
   private static final String REGISTER_PACKAGE = "com.squareup";
-  private static final Uri OAUTH_URI = Uri.parse(
-      "https://connect.squareup.com/oauth2/authorize?client_id="
-          + BuildConfig.CLIENT_ID
-          + "&amp;scope=PAYMENTS_WRITE");
-  private static final Uri OAUTH_DOC_URI =
-      Uri.parse("https://docs.connect.squareup.com/api/oauth/#navsection-oauth");
 
   private final MainActivity activity;
   private final RegisterClient registerClient;
@@ -26,30 +20,6 @@ public class DialogComposer {
   DialogComposer(MainActivity activity, RegisterClient registerClient) {
     this.activity = activity;
     this.registerClient = registerClient;
-  }
-
-  public void showOAuthAlertDialog() {
-    new AlertDialog.Builder(activity).setTitle(R.string.error_not_authorized)
-        .setMessage(activity.getResources().getString(R.string.error_oauth_message))
-        .setPositiveButton(R.string.learn_more, new DialogInterface.OnClickListener() {
-          @Override public void onClick(DialogInterface dialogInterface, int i) {
-            // If the user clicks "Learn More", direct them to the OAuth documentation on
-            // connect.squareup.com
-            Intent oAuthDocIntent = new Intent(Intent.ACTION_VIEW, OAUTH_DOC_URI);
-            oAuthDocIntent.addFlags(Intent.FLAG_ACTIVITY_NEW_TASK);
-            activity.startActivity(oAuthDocIntent);
-          }
-        })
-        .setNegativeButton(R.string.start_oauth, new DialogInterface.OnClickListener() {
-          @Override public void onClick(DialogInterface dialogInterface, int i) {
-            // If the user wants a token now, give them the OAuth link with a redirect URI that
-            // goes to a GitHub page with more information.
-            Intent oAuthIntent = new Intent(Intent.ACTION_VIEW, OAUTH_URI);
-            activity.startActivity(oAuthIntent);
-          }
-        })
-        .setNeutralButton(R.string.ok, null)
-        .show();
   }
 
   public void showErrorDialog(int titleResId, int messageResId) {