+++ CHANGELOG.md
@@ -1,6 +1,15 @@
 Change Log
 ==========
 
+## Version 1.2
+
+_2016-11-2_
+
+* Supported in **Register 4.52 and above.**
+  * Prior to v4.52, Square Register will return an `UNSUPPORTED_API_VERSION` error.
+  * Prior to v4.48, Square Register will return an `INVALID_REQUEST` error
+* Deprecated the `UNAUTHORIZED_CLIENT_ID` error type. OAuth is no longer required for Register API for Android.
+
 ## Version 1.1
 
 _2016-09-21_